+++ Classes/Model/Type/MediaGallery.php
@@ -1,0 +1,25 @@
+<?php
+declare(strict_types=1);
+
+namespace Brotkrueml\Schema\Model\Type;
+
+/*
+ * This file is part of the "schema" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+use Brotkrueml\Schema\Core\Model\AbstractType;
+use Brotkrueml\Schema\Core\Model\WebPageTypeInterface;
+use Brotkrueml\Schema\Model\TypeTrait;
+
+/**
+ * Web page type: Media gallery page. A mixed-media page that can contains media such as images, videos, and other multimedia.
+ */
+final class MediaGallery extends AbstractType implements WebPageTypeInterface
+{
+    use TypeTrait\CreativeWorkTrait;
+    use TypeTrait\ThingTrait;
+    use TypeTrait\WebPageTrait;
+}