+++ Classes/ViewHelpers/Type/MediaGalleryViewHelper.php
@@ -1,0 +1,20 @@
+<?php
+declare(strict_types=1);
+
+namespace Brotkrueml\Schema\ViewHelpers\Type;
+
+/*
+ * This file is part of the "schema" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+use Brotkrueml\Schema\Core\ViewHelpers\AbstractTypeViewHelper;
+
+/**
+ * Web page type: Media gallery page. A mixed-media page that can contains media such as images, videos, and other multimedia.
+ */
+final class MediaGalleryViewHelper extends AbstractTypeViewHelper
+{
+}